Refactor structure accordingly to @mickel8 suggestions
